Refactor C++ back-end into separate modules for finite differencing, CppAD, CodeGen
