+++ include/autogen/autogen.hpp
@@ -1,103 +1,8 @@
 #pragma once
 
-#include <cppad/cg.hpp>
-#include <cppad/cg/arithmetic.hpp>
-#ifdef USE_EIGEN
-#include <cppad/cg/support/cppadcg_eigen.hpp>
-#endif
-
-// clang-format off
-#include "utils/conditionals.hpp"
-#include "cuda/cuda_codegen.hpp"
-#include "cuda/cuda_library_processor.hpp"
-#include "cuda/cuda_library.hpp"
-#include "cg/loop_fun_bridge.hpp"
-#include <cstdio>
-#include <iostream>
-#include <memory>
-#include <stdexcept>
-#include <string>
-#include <array>
-#include <mutex>
-#include <thread>
-// clang-format on
+#include "core/base.hpp"
 
 namespace autogen {
-template <typename BaseScalar>
-using ADCG = typename CppAD::AD<CppAD::cg::CG<BaseScalar>>;
-template <typename BaseScalar>
-using ADFunctor = typename std::function<void(
-    const std::vector<ADCG<BaseScalar>> &, std::vector<ADCG<BaseScalar>> &)>;
-
-template <typename BaseScalar>
-struct FunctionTrace {
-  using CGScalar = typename CppAD::cg::CG<BaseScalar>;
-  using ADCGScalar = typename CppAD::AD<CGScalar>;
-  using ADFun = typename CppAD::ADFun<CGScalar>;
-  using FunBridge = typename CppAD::cg::LoopFunBridge<BaseScalar>;
-
-  std::string name;
-
-  std::shared_ptr<ADFun> tape{nullptr};
-  // TODO fix memory leak here (bridge cannot be destructed without triggering
-  // atomic_index exception in debug mode)
-  FunBridge *bridge{nullptr};
-
-  std::vector<BaseScalar> trace_input;
-
-  ADFunctor<BaseScalar> functor;
-  size_t input_dim;
-  size_t output_dim;
-  std::vector<ADCGScalar> ax;
-  std::vector<ADCGScalar> ay;
-
-  size_t num_iterations;
-  size_t const_input_dim;
-  size_t loop_dependent_dim;
-
-  virtual ~FunctionTrace() {
-    //  delete tape;
-    //  // TODO atomic_base destructor in CppAD gets called twice and throws an
-    //  // exception
-#ifdef NDEBUG
-    // we don't trigger the CPPAD_ASSERT here
-    // delete bridge;
-#endif
-  }
-};
-
-template <typename BaseScalar = double>
-struct CodeGenData {
-  /**
-   * Maps each function that is executed to its trace.
-   */
-  static inline std::map<std::string, FunctionTrace<BaseScalar>> traces;
-  /**
-   * Keeps track of the order of atomic function invocations, i.e. functions
-   * that are called later are added later to this list.
-   */
-  static inline std::vector<std::string> invocation_order;
-
-  /**
-   * Defines whether the current atomic function should record the gradient tape
-   * or not.
-   */
-  static inline bool is_dry_run{true};
-
-  /**
-   * Maps name of the caller to the names of the (atomic) functions it executes.
-   */
-  static inline std::map<std::string, std::vector<std::string>> call_hierarchy;
-
-  static void clear() {
-    traces.clear();
-    invocation_order.clear();
-    call_hierarchy.clear();
-  }
-
-  CodeGenData() = delete;
-};
-
 enum GenerationMode {
   GENERATE_NONE,
   GENERATE_CPPAD,
@@ -125,33 +30,11 @@ static inline std::ostream &operator<<(std::ostream &os,
   return os;
 }
 
-template <template <typename> typename Functor, typename BaseScalar = double>
+template <template <typename> typename Functor>
 struct Generated {
   static inline std::map<std::string, FunctionTrace<BaseScalar>> traces;
 
   const std::string name;
-
-  using CppADScalar = typename CppAD::AD<BaseScalar>;
-  using CGScalar = typename CppAD::AD<CppAD::cg::CG<BaseScalar>>;
-  using ADFun = typename FunctionTrace<BaseScalar>::ADFun;
-  using FunBridge = typename FunctionTrace<BaseScalar>::FunBridge;
-
-  typedef std::unique_ptr<CppAD::cg::GenericModel<BaseScalar>> GenericModelPtr;
-
-  bool compile_in_background{false};
-
-  int num_gpu_threads_per_block{32};
-
-  /**
-   * Step size to use for finite differencing.
-   */
-  double finite_diff_eps{1e-6};
-
-  /**
-   * Whether the generated code is compiled in debug mode (only applies to CPU
-   * and CUDA).
-   */
-  bool debug_mode{false};
 
  protected:
   std::unique_ptr<Functor<BaseScalar>> f_double_;
@@ -159,22 +42,6 @@ struct Generated {
   std::unique_ptr<Functor<CGScalar>> f_cg_;
 
   GenerationMode mode_{GENERATE_CPU};
-  AccumulationMethod jac_acc_method_{ACCUMULATE_NONE};
-
-  mutable std::mutex compilation_mutex_;
-  bool is_compiling_{false};
-
-  // name of the compiled library
-  std::string library_name_;
-
-  size_t local_input_dim_{0};
-  size_t global_input_dim_{0};
-  size_t output_dim_{0};
-
-  // used by CppAD (only in GENERATE_CPPAD mode)
-  std::shared_ptr<CppAD::ADFun<BaseScalar>> tape_{nullptr};
-  std::vector<CppADScalar> ax_;
-  std::vector<CppADScalar> ay_;
 
  public:
   template <typename... Args>
@@ -186,12 +53,7 @@ struct Generated {
     f_cg_ = std::make_unique<Functor<CGScalar>>(std::forward<Args>(args)...);
   }
 
-  // discards the compiled library (so that it gets recompiled at the next
-  // evaluation)
-  void discard_library() {
-    std::lock_guard<std::mutex> guard(compilation_mutex_);
-    library_name_ = "";
-  }
+  
 
   GenerationMode mode() const { return mode_; }
   void set_mode(GenerationMode mode) {
@@ -524,298 +386,8 @@ struct Generated {
     conditionally_compile(compilation_input, outputs[0]);
   }
 
-  FunctionTrace<BaseScalar> trace(const std::vector<BaseScalar> &input,
-                                  std::vector<BaseScalar> &output) const {
-    CodeGenData<BaseScalar>::clear();
-
-    // first, a "dry run" to discover the atomic functions
-    {
-      CodeGenData<BaseScalar>::is_dry_run = true;
-      std::vector<CGScalar> ax(input.size()), ay(output.size());
-      for (size_t i = 0; i < input.size(); ++i) {
-        ax[i] = CGScalar(input[i]);
-      }
-      (*f_cg_)(ax, ay);
-      CodeGenData<BaseScalar>::is_dry_run = false;
-    }
-
-    // next, trace the inner atomic functions
-    const auto &order = CodeGenData<BaseScalar>::invocation_order;
-    for (auto it = order.rbegin(); it != order.rend(); ++it) {
-      FunctionTrace<BaseScalar> &trace = CodeGenData<BaseScalar>::traces[*it];
-      std::cout << "Tracing function \"" << trace.name
-                << "\" for code generation...\n";
-      trace.ax.resize(trace.input_dim);
-      trace.ay.resize(trace.output_dim);
-      for (size_t i = 0; i < trace.input_dim; ++i) {
-        trace.ax[i] = CGScalar(trace.trace_input[i]);
-      }
-      CppAD::Independent(trace.ax);
-      trace.functor(trace.ax, trace.ay);
-      trace.tape = std::make_shared<ADFun>();
-      trace.tape->Dependent(trace.ax, trace.ay);
-      trace.bridge =
-          new FunBridge(trace.name, *(trace.tape), trace.num_iterations,
-                        trace.const_input_dim, trace.loop_dependent_dim);
-    }
-
-    // finally, trace the top-level function
-    FunctionTrace<BaseScalar> trace;
-    trace.name = name;
-    std::vector<CGScalar> ax(input.size()), ay(output.size());
-    std::cout << "Tracing function \"" << name << "\" for code generation...\n";
-    for (size_t i = 0; i < input.size(); ++i) {
-      ax[i] = CGScalar(input[i]);
-    }
-    CppAD::Independent(ax);
-    (*f_cg_)(ax, ay);
-    trace.tape = std::make_shared<ADFun>();
-    trace.tape->Dependent(ax, ay);
-    trace.bridge =
-        new FunBridge(name, *(trace.tape), trace.num_iterations,
-                      trace.const_input_dim, trace.loop_dependent_dim);
-    trace.input_dim = input.size();
-    trace.output_dim = output.size();
-    return trace;
-  }
-
-  void compile_cpu(const FunctionTrace<BaseScalar> &main_trace) {
-    using namespace CppAD;
-    using namespace CppAD::cg;
-
-    ModelCSourceGen<BaseScalar> main_source_gen(*(main_trace.tape), name);
-    main_source_gen.setCreateJacobian(true);
-    ModelLibraryCSourceGen<BaseScalar> libcgen(main_source_gen);
-    // reverse order of invocation to first generate code for innermost
-    // functions
-    const auto &order = CodeGenData<BaseScalar>::invocation_order;
-    std::list<ModelCSourceGen<BaseScalar> *> models;
-    for (auto it = order.rbegin(); it != order.rend(); ++it) {
-      FunctionTrace<BaseScalar> &trace = CodeGenData<BaseScalar>::traces[*it];
-      auto *source_gen = new ModelCSourceGen<BaseScalar>(*(trace.tape), *it);
-      // source_gen->setCreateSparseJacobian(true);
-      // source_gen->setCreateJacobian(true);
-      source_gen->setCreateForwardOne(true);
-      source_gen->setCreateReverseOne(true);
-      models.push_back(source_gen);
-      // we need a stable reference
-      libcgen.addModel(*(models.back()));
-    }
-    libcgen.setVerbose(true);
-#if CPPAD_CG_SYSTEM_WIN
-    SaveFilesModelLibraryProcessor<double> psave(libcgen);
-    psave.saveSourcesTo(name + "_cpu_srcs");
-    std::cerr << "CPU code compilation is not yet available on Windows. Saved "
-                 "source files to \""
-              << name << "_cpu_srcs"
-              << "\".\n ";
-    std::exit(1);
-#else
-    DynamicModelLibraryProcessor<BaseScalar> p(libcgen);
-    auto compiler = std::make_unique<ClangCompiler<BaseScalar>>();
-    compiler->setSourcesFolder(name + "_cpu_srcs");
-    compiler->setSaveToDiskFirst(true);
-    compiler->addCompileFlag("-O" + std::to_string(1));
-    if (debug_mode) {
-      compiler->addCompileFlag("-g");
-      compiler->addCompileFlag("-O0");
-    }
-    p.setLibraryName(name + "_cpu");
-    p.createDynamicLibrary(*compiler, false);
-
-    std::lock_guard<std::mutex> guard(compilation_mutex_);
-    library_name_ = "./" + name + "_cpu.so";
-#endif
-  }
-
-#if !CPPAD_CG_SYSTEM_WIN
-  GenericModelPtr &get_cpu_model() const {
-    static thread_local bool initialized = false;
-    static thread_local auto lib =
-        std::make_unique<CppAD::cg::LinuxDynamicLib<BaseScalar>>(library_name_);
-    static thread_local std::map<std::string, GenericModelPtr> models;
-    if (!initialized) {
-      // load and wire up atomic functions in this library
-      const auto &order = CodeGenData<BaseScalar>::invocation_order;
-      const auto &hierarchy = CodeGenData<BaseScalar>::call_hierarchy;
-      models[name] = GenericModelPtr(lib->model(name).release());
-      for (const std::string &model_name : order) {
-        // we have to keep the atomic function pointers alive
-        models[model_name] = GenericModelPtr(lib->model(model_name).release());
-        // simply add every atomic to the top-level function
-        // (because we don't have hierarchy information for the top-level
-        // function)
-        models[name]->addAtomicFunction(models[model_name]->asAtomic());
-      }
-      for (const auto &[outer, inner_models] : hierarchy) {
-        auto &outer_model = models[outer];
-        for (const std::string &inner : inner_models) {
-          auto &inner_model = models[inner];
-          outer_model->addAtomicFunction(inner_model->asAtomic());
-          std::cout << "Connected atomic function \"" + inner +
-                           "\" to its parent function \""
-                    << outer << "\".\n";
-        }
-      }
-
-      std::cout << "Loaded compiled model \"" << name << "\" from \""
-                << library_name_ << "\".\n";
-      initialized = true;
-    }
-    static thread_local GenericModelPtr &model = models[name];
-    return model;
-  }
-#endif
-
-  void compile_cuda(const FunctionTrace<BaseScalar> &main_trace) {
-    using namespace CppAD;
-    using namespace CppAD::cg;
-
-    CudaModelSourceGen<BaseScalar> main_source_gen(*(main_trace.tape), name);
-    main_source_gen.setCreateJacobian(true);
-    main_source_gen.global_input_dim() = global_input_dim_;
-    main_source_gen.jacobian_acc_method() = jac_acc_method_;
-    CudaLibraryProcessor<BaseScalar> cuda_proc(&main_source_gen,
-                                               name + "_cuda");
-    // reverse order of invocation to first generate code for innermost
-    // functions
-    const auto &order = CodeGenData<BaseScalar>::invocation_order;
-    std::list<CudaModelSourceGen<BaseScalar> *> models;
-    for (auto it = order.rbegin(); it != order.rend(); ++it) {
-      std::cout << "Adding cuda model " << *it << "\n";
-      FunctionTrace<BaseScalar> &trace = CodeGenData<BaseScalar>::traces[*it];
-      auto *source_gen = new CudaModelSourceGen<BaseScalar>(*(trace.tape), *it);
-      source_gen->setCreateForwardOne(true);
-      source_gen->setCreateReverseOne(true);
-      source_gen->set_kernel_only(true);
-      models.push_back(source_gen);
-      cuda_proc.add_model(models.back(), false);
-    }
-    cuda_proc.debug_mode() = debug_mode;
-    cuda_proc.generate_code();
-    cuda_proc.save_sources();
-    cuda_proc.create_library();
-
-    std::lock_guard<std::mutex> guard(compilation_mutex_);
-    library_name_ = name + "_cuda";
-
-    for (auto *model : models) {
-      delete model;
-    }
-  }
-
-  const CudaModel<BaseScalar> &get_cuda_model() const {
-    static thread_local const CudaLibrary<BaseScalar> library(library_name_);
-    static thread_local const CudaModel<BaseScalar> &model =
-        library.get_model(name);
-    return model;
-  }
+  
 };
 
-template <typename BaseScalar = double>
-inline void call_atomic(const std::string &name, ADFunctor<BaseScalar> functor,
-                        const std::vector<ADCG<BaseScalar>> &input,
-                        std::vector<ADCG<BaseScalar>> &output,
-                        size_t num_iterations = 1, size_t const_input_dim = 0,
-                        size_t loop_dependent_dim = 0) {
-  auto &traces = CodeGenData<BaseScalar>::traces;
-  assert(num_iterations > 0);
-  // we only trace for the input for 1 loop iteration
-  size_t actual_input_dim =
-      input.size() - (long(num_iterations) - 1) * loop_dependent_dim;
-
-  if (traces.find(name) == traces.end()) {
-    auto &order = CodeGenData<BaseScalar>::invocation_order;
-    if (!order.empty()) {
-      // the current function is called by another function, hence update the
-      // call hierarchy
-      const std::string &parent = order.back();
-      auto &hierarchy = CodeGenData<BaseScalar>::call_hierarchy;
-      if (hierarchy.find(parent) == hierarchy.end()) {
-        hierarchy[parent] = std::vector<std::string>();
-      }
-      hierarchy[parent].push_back(name);
-    }
-    order.push_back(name);
-    FunctionTrace<BaseScalar> trace;
-    trace.name = name;
-    trace.functor = functor;
-    // only trace for one loop-dependent input (1 iteration)
-    trace.trace_input.resize(actual_input_dim);
-    for (size_t i = 0; i < actual_input_dim; ++i) {
-      BaseScalar raw = CppAD::Value(CppAD::Var2Par(input[i])).getValue();
-      trace.trace_input[i] = raw;
-    }
-    trace.input_dim = actual_input_dim;
-    trace.output_dim = output.size();
-    trace.num_iterations = num_iterations;
-    trace.const_input_dim = const_input_dim;
-    trace.loop_dependent_dim = loop_dependent_dim;
-    traces[name] = trace;
-    functor(input, output);
-    return;
-  } else if (CodeGenData<BaseScalar>::is_dry_run) {
-    // we already preprocessed this function during the dry run
-    return;
-  }
-
-  FunctionTrace<BaseScalar> &trace = traces[name];
-  assert(trace.bridge);
-  // we can only evaluate the functor for one iteration here
-  std::vector<ADCG<BaseScalar>> actual_input(input.begin(),
-                                             input.begin() + actual_input_dim);
-  (*(trace.bridge))(actual_input, output);
-}
-
-template <typename Scalar>
-inline void call_atomic(
-    const std::string &name,
-    const std::function<void(const std::vector<Scalar> &,
-                             std::vector<Scalar> &)> &functor,
-    std::vector<Scalar> input, std::vector<Scalar> &output,
-    size_t num_iterations = 1, size_t const_input_dim = 0,
-    size_t loop_dependent_dim = 0) {
-  // no tracing occurs since the arguments are of type double
-  if (num_iterations > 1) {
-    const size_t m = output.size() + const_input_dim;
-    for (size_t i = 0; i < num_iterations; ++i) {
-      for (size_t j = 0; j < loop_dependent_dim; ++j) {
-        input[m + j] = input[m + i * loop_dependent_dim + j];
-      }
-      functor(input, output);
-      for (size_t j = 0; j < output.size(); ++j) {
-        input[j] = output[j];
-      }
-    }
-  } else {
-    functor(input, output);
-  }
-}
-
-/**
- * More overloads for the atomic function to be traced:
- */
-
-template <typename BaseScalar>
-inline ADCG<BaseScalar> call_atomic(
-    const std::string &name,
-    const std::function<ADCG<BaseScalar>(const std::vector<ADCG<BaseScalar>> &)>
-        &functor,
-    const std::vector<ADCG<BaseScalar>> &input) {
-  ADFunctor<BaseScalar> vec_fun =
-      [functor](const std::vector<ADCG<BaseScalar>> &in,
-                std::vector<ADCG<BaseScalar>> &out) { out[0] = functor(in); };
-  std::vector<ADCG<BaseScalar>> output(1);
-  call_atomic<BaseScalar>(name, vec_fun, input, output);
-  return output[0];
-}
-
-template <typename Scalar>
-inline Scalar call_atomic(
-    const std::string &name,
-    const std::function<Scalar(const std::vector<Scalar> &)> &functor,
-    const std::vector<Scalar> &input) {
-  return functor(input);
-}
 
 }  // namespace autogen
